end_to_end/TPU/gnmt-research-TF-tpu-v4-16/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
chex==0.1.0
colorama==0.4.4
cvxpy==1.1.15
cycler==0.11.0
Cython==0.29.26
dm-tree==0.1.6
ecos==2.0.8
flatbuffers==1.12
flax==0.3.4
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphsurgeon==0.4.5
grpcio==1.43.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.10.0
jax==0.2.19
jaxlib==0.1.70
keras-nightly==2.6.0.dev2021062500
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==11.1.0
libtpu-nightly==0.1.dev20210809
Markdown==3.3.6
matplotlib==3.5.1
msgpack==1.0.3
networkx==2.6.3
numpy==1.21.5
oauthlib==3.1.1
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
optax==0.0.9
osqp==0.6.2.post0
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
plumber-analysis-mkuchnik
portalocker==2.3.2
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools @ git+https://github.com/cocodataset/cocoapi@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
pydot==1.4.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
qdldl==0.1.5.post0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacrebleu==2.0.0
scipy==1.7.3
scs==3.0.0
six==1.15.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
toolz==0.11.2
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
